@@ -1,29 +1,21 @@
 <!-- Navbar de la page Jeux et page_mondes -->
 <template>
     <div class="home_banner">
-      <img alt="Menu" src='@/assets/menu/burger-bar.png'/>
+      <img alt="Menu" src='/menu/burger-bar.png'/>
       <h1>{{ title }}</h1>
       <div class="image-container">
-        <img class='avatar-icon' alt="Avatar choisi" src='@/assets/players/player1.png'/>
+        <img class='avatar-icon' alt="Avatar choisi" src="/players/player1.png"/>
       </div>
     </div>
   </template>
   
-  <script lang="ts">
-  import { Options, Vue } from 'vue-class-component';
-  import BannerMenu from './BannerMenu.vue';
+  <script setup lang="ts">
+
+  const props = defineProps({
+    title: String
+  });
   
-  @Options({
-    props: {
-      title: String
-    },
-    components: {
-      BannerMenu
-    }
-  })
-  export default class HomeBanner extends Vue {
-    title!: string
-  }
+  const title =''
   </script>
   
   <style scoped>
