@@ -1,28 +1,22 @@
 <template>
   <div class="star-rating">
-    <img v-for="star in rating" :key="star" :src="getImagePath(star)" class="star-img" :alt="title" />
+    <img v-for="star in rating" :key="star" :src="'/'+world_num+'/castle'+star+'.png'" class="star-img" :alt="title" />
   </div>
 </template>
 
-<script lang="ts">
-  import { Options, Vue } from 'vue-class-component';
+<script setup lang="ts">
 
-  @Options({
-    props: {
-      rating: Number,
+  const props = defineProps({
+    rating: Number,
       world_num: String,
       title: String
-    }
-  })
-  export default class StarRating extends Vue {
-    rating!: number;
-    world_num!: string;
-    title!: string;
+  });
 
-    getImagePath(star: number) {
-      return require(`@/assets/${this.world_num}/castle${star}.png`);
-    }
-  }
+  
+  const rating =0;
+  const world_num = 'world1';
+  const title='';
+
 </script>
 
 <style scoped>
