@@ -1,23 +1,21 @@
-import { createRouter, createWebHistory, RouteRecordRaw } from 'vue-router'
+import { createRouter, createWebHistory } from 'vue-router'
 import GamesView from '../views/GamesView.vue'
 import SafeCityView from '../views/SafeCityView.vue'
 
-const routes: Array<RouteRecordRaw> = [
-  {
-    path: '/',
-    name: 'home',
-    component: GamesView
-  },
-  {
-    path: '/safecity',
-    name: 'safe-city',
-    component: SafeCityView
-  }
-]
-
 const router = createRouter({
-  history: createWebHistory(process.env.BASE_URL),
-  routes
+  history: createWebHistory(import.meta.env.BASE_URL),
+  routes: [
+    {
+      path: '/',
+      name: 'home',
+      component: GamesView
+    },
+    {
+      path: '/safecity',
+      name: 'safe-city',
+      component: SafeCityView
+    }
+  ]
 })
 
 export default router
